fuse - cli
