--- perceiverio/cli/__init__.py
@@ -1,1 +1,0 @@
-from perceiverio.cli.cli import CLI
